pyproject.toml
[tool.poetry]
name = "sygnal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
pandas = "^1.4.0"
numpy = "^1.22.2"
ta = "^0.9.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-cors-headers = "^3.11.0"
djangorestframework-simplejwt = "^5.0.0"
discord = "^1.7.3"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
